@@ -12,17 +12,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^9.1.12",
+    "@angular/animations": "^10.0.6",
     "@angular/cdk": "^9.2.4",
-    "@angular/common": "~9.1.12",
-    "@angular/compiler": "~9.1.12",
-    "@angular/core": "~9.1.12",
-    "@angular/forms": "~9.1.12",
+    "@angular/common": "~10.0.6",
+    "@angular/compiler": "~10.0.6",
+    "@angular/core": "~10.0.6",
+    "@angular/forms": "~10.0.6",
     "@angular/material": "^9.2.4",
     "@angular/material-moment-adapter": "^9.2.4",
-    "@angular/platform-browser": "~9.1.12",
-    "@angular/platform-browser-dynamic": "~9.1.12",
-    "@angular/router": "~9.1.12",
+    "@angular/platform-browser": "~10.0.6",
+    "@angular/platform-browser-dynamic": "~10.0.6",
+    "@angular/router": "~10.0.6",
     "core-js": "^2.6.2",
     "font-awesome": "^4.7.0",
     "material-design-icons": "^3.0.1",
@@ -35,8 +35,8 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.901.12",
     "@angular/cli": "^9.1.12",
-    "@angular/compiler-cli": "~9.1.12",
-    "@angular/language-service": "~9.1.12",
+    "@angular/compiler-cli": "~10.0.6",
+    "@angular/language-service": "~10.0.6",
     "@types/jasmine": "~3.3.5",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
@@ -51,6 +51,6 @@
     "protractor": "~5.4.2",
     "ts-node": "~7.0.0",
     "tslint": "~5.12.1",
-    "typescript": "3.8.3"
+    "typescript": "3.9.7"
   }
 }
